Fixed: rename and fix packages controller, service, entity
